Split logic on two modules
