+++ salesforce-driver/docs/birt/Data source - property binding.png
