+++ salesforce-driver/src/main/java/com/ascendix/jdbc/salesforce/connection/ForceConnectionInfo.java
