+++ salesforce-driver/src/main/java/com/ascendix/jdbc/salesforce/metadata/Column.java
