+++ salesforce-driver/src/main/java/com/ascendix/jdbc/salesforce/metadata/ColumnMap.java
