+++ salesforce-driver/src/main/java/com/ascendix/jdbc/salesforce/oauth/ForceOAuthClient.java
