+++ salesforce-driver/src/main/java/com/ascendix/jdbc/salesforce/oauth/ForceUserInfo.java
@@ -1,6 +1,7 @@
 package com.ascendix.jdbc.salesforce.oauth;
 
 import com.google.api.client.util.Key;
+import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -8,6 +9,7 @@ import java.util.Map;
 
 @Data
 @NoArgsConstructor
+@AllArgsConstructor
 public class ForceUserInfo {
 
     @Key("user_id")
