+++ salesforce-driver/src/main/java/com/ascendix/jdbc/salesforce/resultset/CachedResultSet.java
