+++ salesforce-driver/src/main/java/com/ascendix/jdbc/salesforce/resultset/CachedResultSetMetaData.java
