+++ salesforce-driver/src/main/java/com/ascendix/jdbc/salesforce/statement/ParameterMetadataImpl.java
